Fix the way samples read
This is a pretty major fix. I had thought that there were 70 channels for each of L and R, each 16-bit integers. But in fact there are actually 35 of each, and none at same frequency. PRA_sampling.pdf sketch shows what is happening.

This means that in some sense comparing L[i] to R[i], e.g. calculating L[i]^2+R[i]^2 is meaningless, where i is an array index for frequency. At least it is for emission narrower than channel separation which is 19.2 kHz for VG2 LOWBAND. But if signal is broader band than this it can be meaningful. But I think it makes sense to nominally attribute the frequency midpoint between L[i] and R[i] to any such quantities.

maser4py seems to copy each value so that there are 70 channels for each of L and R, but every pair of values is copied. e.g. channels 0 and 1 are identical, channels 2 and 3 are identical, ... channels 68 and 69 are identical. This seems to be the case for all data I've checked.

Leaving it with 35 channels as this is how many measurements there are.